Move Handler interface into their own package to better illustrate/group together. Remove IFlow marker Merge SimpleProtocolFactory into SimpleConnection
