--- src/main/java/net/minecraftforge/network/simple/BaseFlow.java
@@ -1,9 +1,0 @@
-/*
- * Copyright (c) Forge Development LLC and contributors
- * SPDX-License-Identifier: LGPL-2.1-only
- */
-
-package net.minecraftforge.network.simple;
-
-public interface BaseFlow {
-}
