+++ src/main/java/net/minecraftforge/network/simple/SimpleConnection.java
@@ -9,10 +9,24 @@ import io.netty.util.AttributeKey;
 import net.minecraft.network.FriendlyByteBuf;
 import net.minecraft.network.RegistryFriendlyByteBuf;
 import net.minecraftforge.network.NetworkProtocol;
+import net.minecraftforge.network.simple.handler.SimpleHandlerProtocol;
+import net.minecraftforge.network.simple.handler.SimplePacket;
 
 import java.util.function.Consumer;
 
-public interface SimpleConnection<BASE> extends SimpleProtocolFactory {
+public interface SimpleConnection<BASE> {
+    /**
+     * Creates a builder grouping together all packets under the same protocol.
+     * This will validate that the protocol matches before the packet is sent or received.
+     */
+    <NEWBUF extends FriendlyByteBuf, NEWBASE> SimpleProtocol<NEWBUF, NEWBASE> protocol(NetworkProtocol<NEWBUF> protocol);
+
+    /**
+     * Creates a builder grouping together all packets under the same protocol.
+     * This will validate that the protocol matches before the packet is sent or received.
+     */
+    <NEWBUF extends FriendlyByteBuf, CTX, NEWBASE extends SimplePacket<CTX>> SimpleHandlerProtocol<NEWBUF, NEWBASE> protocol(AttributeKey<CTX> context, NetworkProtocol<NEWBUF> protocol);
+
     /**
      * Consumer version of {@link #protocol(NetworkProtocol)}. The Consumer will immediately be called with the created protocol.
      */
